More content and bugs fixes.
Pedestal routing: Ensured that splat routes are move to the end of the list to avoid eclipsing with other routes when routes are moved together.

Examples: Added an example site to show the README.md. This will eventually form the basis of the Jig UI. Bootstrap 3.0.0 assets added.

Firefox reload: Imported the functionality from Up to reload firefox via the Firefox Remote Control Addon.