A fix to the classloading issue whereby protocol dispatch failed to work on subsequent resets.
The problem was caused by the user namespace, and consequently the system map, being unloaded and reloaded. This meant that the project's classloader was lost, and that new ones were created for the reset, meaning that classes in a project were split between two classloaders. The fix required that the user namespace is not reloaded by tools.namespace, and therefore metadata has been applied to it.