Replace classloading with Pomegranate.
The original classloading design of loading the classes for each project in a separate classloader created problems in resolving symbols to source files for nREPL clients and the workaround explained in the README.md was too much of a hack.

In #7, Ivan Kozik asked for clarification about the reasons for using classloader and this led to a re-think. Using Pomegranate to add a project's dependencies to the normal Clojure classloader hierarchy turns out to be a much better approach.

There cannot be true isolation between projects (allowing, for instance, separate projects to have different versions of the same Clojure dependency), because all Clojure symbols are combined in Clojure's namespace maps (and there is no prospect for version isolation of these symbols). So the constraint of non-conflicting dependencies in a Clojure runtime is completely acceptable - developers can get around these problems in the usual ways (conflict resolution, exclusions, etc.)

Fixes #7.